Transition from two-dimensional ballistics calculation to global 3D (linear mathematics and vectors)
Convert State to use Vector3 position/velocity and update all integrators and handlers to the new API. Change integration step parameter from double to float across IIntegrator, Euler, RK4 and RKF45; simplify RKF45 implementation and adaptive step logic; unify termination check and add small ground margin. Rewrite BallisticsHandler derivatives to use Vector3 math (relative velocity, drag, gravity, thrust) with safe mass handling, remove CSV logging and expose Recalculate(Vector3) for explicit initial velocity (muzzle + vehicle). Update Projectile to use precomputed path positions and only start flight when path exists. Improve TrajectoryRenderer performance (array reuse, Awake init, debug vectors) and tidy scene asset: switch default integration method, remove unused components/settings. Overall cleanup and modernization of trajectory/integration pipeline.

+++ BallisticsSimulation/Assets/Ballistics/Scripts/Modern/IIntegrator.cs
@@ -6,7 +6,7 @@ namespace BallisticsSimulation
     {
         public List<State> Calculate(
             in State initState, 
-            double step, 
+            float step, 
             int maxSteps, 
             BallisticsHandler handler,
             double eps = 1e-4,
